added a bunch of new shit, see PR
